Refactor views to separate package
