+++ app/src/main/java/com/example/fitnessapp/navigation/AppNavigation.kt
@@ -1,21 +1,14 @@
 package com.example.fitnessapp.navigation
 
-import android.R.attr.type
 import androidx.compose.runtime.Composable
-import androidx.navigation.NavHostController
 import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import com.example.fitnessapp.CurrentWorkoutScreen
-import com.example.fitnessapp.ExerciseListSelectionScreen
-import com.example.fitnessapp.ExerciseHistoryScreen
-import com.example.fitnessapp.viewmodel.CurrentWorkoutViewModel
-import com.example.fitnessapp.viewmodel.ExerciseHistoryViewModel
-import org.koin.androidx.compose.getViewModel
-import org.koin.androidx.compose.koinViewModel
-import org.koin.core.parameter.parametersOf
+import com.example.fitnessapp.views.CurrentWorkoutScreen
+import com.example.fitnessapp.views.ExerciseListSelectionScreen
+import com.example.fitnessapp.views.ExerciseHistoryScreen
 
 @Composable
 fun AppNavigation() {
