+++ app/src/main/java/com/example/fitnessapp/views/ExerciseListSelectionScreen.kt
@@ -1,4 +1,4 @@
-package com.example.fitnessapp
+package com.example.fitnessapp.views
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
